Break out mock_github_request_handler
