+++ tests/utils/__init__.py
@@ -1,6 +1,5 @@
 """Collection of classes and methods used by various hvac test cases."""
 import base64
-import json
 import logging
 import operator
 import os
@@ -16,12 +15,6 @@ from mock import patch
 
 from hvac import Client
 
-try:
-    # Python 2.7
-    from http.server import BaseHTTPRequestHandler
-except ImportError:
-    # Python 3.x
-    from BaseHTTPServer import BaseHTTPRequestHandler
 
 logger = logging.getLogger(__name__)
 
@@ -388,68 +381,3 @@ class HvacIntegrationTestCase(object):
         :type mount_point: str
         """
         self.client.disable_secret_backend(mount_point)
-
-
-class MockGithubRequestHandler(BaseHTTPRequestHandler):
-    """Small HTTP server used to mock out certain GitHub API routes that vault requests in the github auth method."""
-
-    def do_GET(self):
-        """Dispatch GET requests to associated mock GitHub 'handlers'."""
-        self.send_response(200)
-        self.send_header('Content-Type', 'application/json')
-        self.end_headers()
-
-        if self.path == '/user':
-            self.do_user()
-        elif self.path == '/user/orgs?per_page=100':
-            self.do_organizations_list()
-        elif self.path == '/user/teams?per_page=100':
-            self.do_team_list()
-        return
-
-    def log_message(self, format, *args):
-        """Squelch any HTTP logging."""
-        return
-
-    def do_user(self):
-        """Return the bare minimum GitHub user data needed for Vault's github auth method."""
-        response = {
-            "login": "hvac-dude",
-            "id": 1,
-        }
-
-        self.wfile.write(json.dumps(response).encode())
-
-    def do_organizations_list(self):
-        """Return the bare minimum GitHub organization data needed for Vault's github auth method.
-
-        Only returns data if the request Authorization header has a contrived github token value of "valid-token".
-        """
-        response = []
-        if self.headers.get('Authorization') == 'Bearer valid-token':
-            response.append(
-                {
-                    "login": "hvac",
-                    "id": 1,
-                }
-            )
-
-            self.wfile.write(json.dumps(response).encode())
-
-    def do_team_list(self):
-        """Return the bare minimum GitHub team data needed for Vault's github auth method.
-
-        Only returns data if the request Authorization header has a contrived github token value of "valid-token".
-        """
-        response = []
-        if self.headers.get('Authorization') == 'Bearer valid-token':
-            response.append(
-                {
-                    "name": "hvac-team",
-                    "slug": "hvac-team",
-                    "organization": {
-                        "id": 1,
-                    }
-                }
-            )
-        self.wfile.write(json.dumps(response).encode())
